revert back to using the short date for dob, make the theme's locale yaml file look similar to the pomo locale.yaml with the addition of having translations there as well rather than in a separate (pomo) file

+++ plugins/productGroupPricing/lib/ZMProductGroupPricing.php
@@ -84,8 +84,8 @@ class ZMProductGroupPricing extends ZMObject {
             // default to current date
             $startDate = new DateTime();
         }
-        $this->startDate_ = DateTime::createFromFormat(ZMLocales::instance()->getLocale()->getFormat('date', 'long'), $startDate);
-        $this->endDate_ = DateTime::createFromFormat(ZMLocales::instance()->getLocale()->getFormat('date', 'long'), $request->getParameter('endDate'));
+        $this->startDate_ = DateTime::createFromFormat(ZMLocales::instance()->getLocale()->getFormat('date', 'short'), $startDate);
+        $this->endDate_ = DateTime::createFromFormat(ZMLocales::instance()->getLocale()->getFormat('date', 'short'), $request->getParameter('endDate'));
     }
